// app/jazz/htmlTypes.d.ts
export type EventListenerWithTarget<E extends Event, T> = (e: Omit<E, "target"> & { target: T }) => void;
export type EventsMap<T extends keyof HTMLElementTagNameMap> = Partial<{ [K in keyof HTMLElementEventMap]: EventListenerWithTarget<HTMLElementEventMap[K], HTMLElementTagNameMap[T]> }>;
export type StyleMap = Partial<Record<keyof CSSStyleDeclaration, string | number>>;
export interface HtmlComponentProps<T extends keyof HTMLElementTagNameMap> extends JSX.IntrinsicAttributes {
  class?: string;
  className?: string;
  events?: EventsMap<T>;
  style?: StyleMap;
  id?: string;
  ref?: Ref<HTMLElementTagNameMap[T]>;
}

export interface HtmlContentComponentProps<
  T extends keyof HTMLElementTagNameMap,
  C = UINode | UINode[],
> extends HtmlComponentProps<T> {
  content?: C;
}

export type HtmlInputProps = HtmlComponentProps<"input"> & {
  value?: string;
};

export type HtmlAProps = HtmlContentComponentProps<"a"> & {
  href?: string;
  target?: string;
};
// ...
    interface IntrinsicAttributes {
      key?: any;
    }